refactor: add wind module, add relevant items to it
