--- src/bundles.rs
@@ -1,11 +1,0 @@
-use crate::gameplay::components::*;
-use bevy::prelude::*;
-use bevy_ecs_ldtk::prelude::*;
-
-#[derive(Clone, Bundle, LdtkEntity)]
-pub struct GrassBundle {
-    #[sprite_sheet_bundle]
-    #[bundle]
-    pub sprite_sheet_bundle: SpriteSheetBundle,
-    pub wind_timer: WindTimer,
-}
