+++ src/main.rs
@@ -2,7 +2,6 @@
 #![allow(clippy::too_many_arguments, clippy::type_complexity)]
 
 mod animation;
-mod bundles;
 mod camera;
 mod control_display;
 mod event_scheduler;
@@ -22,6 +21,7 @@ mod sokoban;
 mod sugar;
 mod ui;
 mod willo;
+mod wind;
 
 use animation::SpriteSheetAnimationPlugin;
 use bevy::{prelude::*, render::texture::ImageSettings};
@@ -30,7 +30,6 @@ use bevy_asset_loader::prelude::*;
 use bevy_easings::EasingsPlugin;
 use bevy_ecs_ldtk::prelude::*;
 use iyes_loopless::prelude::*;
-use rand::Rng;
 
 pub const UNIT_LENGTH: f32 = 32.;
 
@@ -106,8 +105,8 @@ fn main() {
                 .run_in_state(GameState::Gameplay)
                 .label(history::FlushHistoryCommands),
         )
-        .add_system(sugar::animate_grass_system.run_not_in_state(GameState::AssetLoading))
-        .register_ldtk_entity::<bundles::GrassBundle>("Grass");
+        .add_system(wind::animate_grass_system.run_not_in_state(GameState::AssetLoading))
+        .register_ldtk_entity::<wind::GrassBundle>("Grass");
 
     #[cfg(feature = "hot")]
     {
